Stabilizing to be more package-like.
