+++ gbasm/__init__.py
@@ -1,27 +1,28 @@
 # Game Boy Assembler
 
-import imp
-try:
-    imp.find_module('gbasm_dev')
-    from gbasm_dev import set_gbasm_path
-    set_gbasm_path()
-except ImportError:
-    pass
+# import imp
+# try:
+#     imp.find_module('gbasm_dev')
+#     from gbasm_dev import set_gbasm_path
+#     set_gbasm_path()
+# except ImportError:
+#     pass
 
-from gbasm.reader import Reader, BufferReader, FileReader
-from gbasm.exception import Error, ErrorCode
-from gbasm.instruction import Instruction, InstructionSet, InstructionPointer
-from gbasm.instruction.registers import Registers
-from gbasm.section import Section, SectionType, SectionAddress
-from gbasm.equate import Equate
-from gbasm.storage import StorageType, Storage, StorageParser
-from gbasm.conversions import ExpressionConversion
-from gbasm.label import Label, Labels
-from gbasm.assembler import Assembler
-import gbasm.resolver
+# from . import core
+# from .core import instruction
+# from core.instruction import *
+# from assembler import *
 
-__all__ = [
-    "reader", "exception", "basic_lexer", "equate", "constants",
-    "conversions", "label", "resolver", "section", "storage",
-    "assembler"
-]
+from .core import Reader, BufferReader, FileReader
+# from constants import NODE, DIR, TOK, EQU, LBL, INST, STOR, SEC
+# from gbasm.exception import Error, ErrorCode, ParserException
+# from instruction import LexerResults, Instruction, Registers, \
+#      InstructionPointer, InstructionSet, LR35902Data
+# from gbasm.section import Section, SectionType, SectionAddress
+# from gbasm.equate import Equate
+# from gbasm.storage import StorageType, Storage, StorageParser
+# from gbasm.node_processor import CodeNode, NodeProcessor
+# from gbasm.label import Label, Labels
+# from gbasm.basic_lexer import BasicLexer, is_compound_node, is_node_valid
+# from gbasm.resolver import Resolver
+# from gbasm.node_processor import CodeNode, NodeProcessor
