+++ gbasm/assembler/node_processor.py
@@ -7,28 +7,28 @@ from collections import namedtuple
 import tempfile
 import pprint
 
-import gbasm.core as core
-import gbasm.core.constants as const
-import gbasm.assembler as asm
+from ..core import TOK, DIR, LBL, EQU, INST, SEC, STOR
+from ..core import ExpressionConversion, InstructionSet, InstructionPointer
+from ..core import Reader, BasicLexer, Section, Equate, Label, Labels, Storage
+from ..core import ErrorCode, Error, Instruction, ParserException
+from ..core import is_compound_node, is_node_valid
+from ..core import NodeType
+from .code_node import CodeNode, CodeOffset
 
-EC = core.ExpressionConversion
-IS = core.InstructionSet
-IP = core.InstructionPointer
-
-TOK = const.TOK
-DIR = const.DIR
-LBL = const.LBL
+EC = ExpressionConversion
+IS = InstructionSet
+IP = InstructionPointer
 
 class NodeProcessor(object):
-    def __init__(self, reader: core.Reader):
+    def __init__(self, reader: Reader):
         self._reader = reader
         self._line_no = 0
-        self._lexer = core.BasicLexer(reader)
-        self._code: [asm.CodeNode] = []
+        self._lexer = BasicLexer(reader)
+        self._code: [CodeNode] = []
         self._bad: [dict] = []
-        self._sections: [core.Section] = []
+        self._sections: [Section] = []
 
-    def process_EQU(self, tokens: list) -> asm.CodeNode:
+    def process_EQU(self, tokens: list) -> CodeNode:
         """Process an EQU statement. """
         # The tokens should always be multiple since an EQU contains
         # a label followed by the EQU to associate with the label
@@ -38,28 +38,28 @@ class NodeProcessor(object):
             return None
         if tokens[1][DIR] != 'EQU':
             return None
-        result = core.Equate(tokens)
+        result = Equate(tokens)
         if result:
             result.parse()
-            lbl = core.Label(result.name(), result.value(), constant=True)
-            core.Labels().add(lbl)
-            return asm.CodeNode(const.NodeType.EQU, lbl, IP().offset_from_base())
-        err = core.Error(core.ErrorCode.INVALID_LABEL_NAME,
+            lbl = Label(result.name(), result.value(), constant=True)
+            Labels().add(lbl)
+            return CodeNode(NodeType.EQU, lbl, IP().offset_from_base())
+        err = Error(ErrorCode.INVALID_LABEL_NAME,
                     source_file=self._reader.filename,
                     source_line=int(self._reader.line))
         tokens["error"] = err
         # _errors.append(err)
-        return asm.CodeNode(const.NodeType.NODE, tokens, 0)
+        return CodeNode(NodeType.NODE, tokens, 0)
 
-    def process_INSTRUCTION(self, node: dict) -> asm.CodeNode:
+    def process_INSTRUCTION(self, node: dict) -> CodeNode:
         ins = None
         if node is None:
             return None
-        ins = core.Instruction(node)
+        ins = Instruction(node)
         if ins.parse_result().is_valid():
             offset = IP().offset_from_base()
             IP().move_relative(len(ins.machine_code()))
-            return asm.CodeNode(const.NodeType.INST, ins, offset)
+            return CodeNode(NodeType.INST, ins, offset)
         # Instruction is not valid. This could mean either it really is
         # invalid (typo, wrong argument, etc) or that it has a label. To
         # get started, just check to make sure the mnemonic is at least
@@ -69,38 +69,38 @@ class NodeProcessor(object):
             ins2 = asm.Resolver().resolve_instruction(ins, IP().location)
             if ins2 and ins2.is_valid():
                 IP().move_relative(len(ins2.machine_code()))
-                return asm.CodeNode(const.NodeType.INST, ins2, offset)
+                return CodeNode(NodeType.INST, ins2, offset)
         if ins and ins.is_valid():
             IP().move_relative(len(ins.machine_code()))
-            return asm.CodeNode(const.NodeType.INST, ins, offset)
-        return asm.CodeNode(const.NodeType.NODE, node, offset)  # Error, return the errant node
+            return CodeNode(NodeType.INST, ins, offset)
+        return CodeNode(NodeType.NODE, node, offset)  # Error, return the errant node
 
-    def process_LABEL(self, node: dict, value=None) -> asm.CodeNode:
+    def process_LABEL(self, node: dict, value=None) -> CodeNode:
         if not node:
             return None
         if node[DIR] != LBL:
             return None
         clean = node[TOK].strip("()")
-        existing = core.Labels()[clean]
+        existing = Labels()[clean]
         if existing:
             return None
         loc = value
         if not value:
             loc = IP().location
-        label = core.Label(clean, loc)
-        return asm.CodeNode(const.NodeType.LBL, label, IP().offset_from_base())
+        label = Label(clean, loc)
+        return CodeNode(NodeType.LBL, label, IP().offset_from_base())
 
-    def process_SECTION(self, tokens: dict) -> asm.CodeNode:
-        if not tokens or (tokens and tokens[0] != const.SEC):
+    def process_SECTION(self, tokens: dict) -> CodeNode:
+        if not tokens or (tokens and tokens[0] != SEC):
             return None
         if len(tokens) < 3:
             return None
         try:
             section = self._find_section(' '.join(tokens))
-        except core.ParserException:
+        except ParserException:
             return None
         if section is None:  # not found, create a new one.
-            secn = core.Section(tokens)
+            secn = Section(tokens)
             # print("Processing SECTION")
             self._sections.append(secn)
             num_addr, _ = secn.address_range()
@@ -110,21 +110,21 @@ class NodeProcessor(object):
 
             return secn
 
-    def process_STORAGE(self, node: dict) -> asm.CodeNode:
+    def process_STORAGE(self, node: dict) -> CodeNode:
         if not node:
             return None
         offset = IP().offset_from_base()
-        sto = core.Storage(node)
+        sto = Storage(node)
         # print(f"Processing Storage type {sto.storage_type()}")
         # print(f"Storage len = {len(sto)}")
-        return asm.CodeNode(const.NodeType.STOR, sto, offset)
+        return CodeNode(NodeType.STOR, sto, offset)
 
-    def process_node(self, node: dict) -> [asm.CodeNode]:
-        nodes: [asm.CodeNode] = []
-        if not core.is_node_valid(node):
+    def process_node(self, node: dict) -> [CodeNode]:
+        nodes: [CodeNode] = []
+        if not is_node_valid(node):
             self._bad.append(node)
             return None
-        if core.is_compound_node(node):
+        if is_compound_node(node):
             # The MULTIPLE case is when a LABEL is on the same line as some
             # other data like an instruction. In some cases this is common
             # like an EQU that is supposed to contain both a LABEL and a
@@ -132,42 +132,42 @@ class NodeProcessor(object):
             # INSTRUCTION.
             multi = self.process_compound_node(node)
             if not multi:
-                nodes.append(asm.CodeNode(const.NodeType.NODE, node, IP().offset_from_base()))
+                nodes.append(CodeNode(NodeType.NODE, node, IP().offset_from_base()))
             else:
                 nodes.extend(multi)
             return nodes
-        if node[DIR] == const.SEC:
+        if node[DIR] == SEC:
             sec = self.process_SECTION(node[TOK])
             if sec is None:
                 msg = f"Error in parsing section directive. "\
                     "{self.filename}:{self._line_no}"
-                err = core.Error(core.ErrorCode.INVALID_SECTION_POSITION,
+                err = Error(ErrorCode.INVALID_SECTION_POSITION,
                             supplimental=msg,
                             source_line=self._line_no)
                 node["error"] = err
-                nodes.append(asm.CodeNode(const.NodeType.NODE, node, 0))
+                nodes.append(CodeNode(NodeType.NODE, node, 0))
             else:
                 # address = sec.address_range()
-                # nodes.append(asm.CodeNode(SEC, sec, address.start))
-                nodes.append(asm.CodeNode(const.NodeType.SEC, sec, 0))
+                # nodes.append(CodeNode(SEC, sec, address.start))
+                nodes.append(CodeNode(NodeType.SEC, sec, 0))
             return nodes
         # Just check for a label on it's own line.
         elif node[DIR] == LBL:
             label = self.process_LABEL(node)
-            core.Labels().add(label.code_obj)
+            Labels().add(label.code_obj)
             nodes.append(label)
         # Lastly, if not any of the above, it _might_be an instruction
-        elif node[DIR] == const.INST:
+        elif node[DIR] == INST:
             ins = self.process_INSTRUCTION(node)
             if ins:
                 nodes.append(ins)
             else:
-                nodes.append(asm.CodeNode(const.NodeType.NODE, node, IP().offset_from_base()))
+                nodes.append(CodeNode(NodeType.NODE, node, IP().offset_from_base()))
         return nodes
 
-    def process_compound_node(self, node: dict) -> [asm.CodeNode]:
+    def process_compound_node(self, node: dict) -> [CodeNode]:
         tok_list = node[TOK]
-        nodes: [asm.CodeNode] = []
+        nodes: [CodeNode] = []
         if len(tok_list) < 2:
             # err = Error(ErrorCode.INVALID_DECLARATION,
             #             source_line=self._line_no)
@@ -175,27 +175,27 @@ class NodeProcessor(object):
         # Record a label unless it's an equate. The equate object (which is
         # similar to a label) handles the storage of both.
         if tok_list[0][DIR] == LBL and \
-            tok_list[1][DIR] != const.EQU:
+            tok_list[1][DIR] != EQU:
             clean = tok_list[0][TOK].strip("()")
-            existing = core.Labels()[clean]
+            existing = Labels()[clean]
             if not existing:
                 label = self.process_LABEL(tok_list[0])
-                core.Labels().add(label.code_obj)
+                Labels().add(label.code_obj)
             else:
-                label = asm.CodeNode(const.NodeType.LBL, existing, IP().offset_from_base())
+                label = CodeNode(NodeType.LBL, existing, IP().offset_from_base())
             nodes.append(label)
         # Equate has it's own required label. It's not a standard label
         # in that it can't start with a '.' or end with a ':'
-        if tok_list[1][DIR] == const.EQU:
+        if tok_list[1][DIR] == EQU:
             equ = self.process_EQU(node[TOK])
             nodes.append(equ)
         # An instruction is allowed to be on the same line as a label.
-        elif tok_list[1][DIR] == const.INST:
+        elif tok_list[1][DIR] == INST:
             ins = self.process_INSTRUCTION(tok_list[1])
             nodes.append(ins)
         # Storage values can be associated with a label. The label
         # then can be used almost like an EQU label.
-        elif tok_list[1][DIR] == const.STOR:
+        elif tok_list[1][DIR] == STOR:
             storage = self.process_STORAGE(tok_list[1])
             if storage:
 
@@ -203,17 +203,17 @@ class NodeProcessor(object):
                 nodes.append(storage)
                 return nodes
         else:
-            nodes.append(asm.CodeNode(const.NodeType.NODE, node, IP().offset_from_base()))
+            nodes.append(CodeNode(NodeType.NODE, node, IP().offset_from_base()))
         return nodes
 
-    def _find_section(self, line: str) -> core.Section:
+    def _find_section(self, line: str) -> Section:
         try:
-            section = core.Section(line)
-        except core.ParserException:
+            section = Section(line)
+        except ParserException:
             fname = self._reader.filename()
             msg = f"Parser exception occured {fname}:{self._line_no}"
             print(msg)
-            raise core.ParserException(msg, line_number=self._line_no)
+            raise ParserException(msg, line_number=self._line_no)
         else:
             if section:
                 for _, val in enumerate(self._sections):
