+++ gbasm/core/__init__.py
@@ -3,23 +3,25 @@ from .reader import Reader, BufferReader, FileReader
 from .conversions import ExpressionConversion
 from .constants import NodeType, NODE_TYPES, DIRECTIVES, STORAGE_DIR
 from .constants import NODE, DIR, TOK, EQU, LBL, INST, STOR, SEC, MULT, EXT
+from .label import Label, Labels, LabelUtils, LabelScope
 from .equate import Equate
 
-from .basic_lexer import BasicLexer, is_compound_node
-from .basic_lexer import is_node_valid, tokenize_line
-from .lexer_parser import LexerResults, LexerTokens, InstructionParser
+from .lexer_parser import LexerResults, LexerTokens, InstructionParser,\
+    BasicLexer, is_compound_node, is_node_valid, tokenize_line
 
 from .exception import ParserException, DefineDataError
 from .exception import SectionDeclarationError, SectionTypeError
 from .exception import ErrorCode, Error
 
-from .label import Label, Labels, LabelUtils, LabelScope
 
 from .storage import Storage, StorageType
 from .section import Section, SectionAddress, SectionType
 
-from .instruction import LR35902Data, InstructionSet, Registers
-from .instruction import Instruction, InstructionPointer
+from .lr35902_data import LR35902Data
+from .instruction_set import  InstructionSet
+from .registers import Registers
+from .instruction import Instruction
+from .instruction_pointer import InstructionPointer
 
 __all__ = [
     "Reader", "BufferReader", "FileReader", "ExpressionConversion",
