--- gbasm/core/instruction/__init__.py
@@ -1,9 +1,0 @@
-"""
-Game Boy Assembler Instruction related classes
-"""
-
-from .lr35902_data import LR35902Data
-from .instruction_set import InstructionSet
-from .instruction_pointer import InstructionPointer
-from .registers import Registers
-from .instruction import Instruction
