--- gbasm/core/instruction/lr35902_data.py
@@ -1,729 +1,0 @@
-from singleton_decorator import singleton
-
-@singleton
-class LR35902Data():
-    def __init__(self):
-        self.json = """
-{"$00":{"mnemonic":"NOP","length":1,"cycles":[4],"flags":["-","-","-","-"],"addr\
-":"$00"},"$01":{"mnemonic":"LD","length":3,"cycles":[12],"flags":["-","-","-","-\
-"],"addr":"$01","operand1":"BC","operand2":"d16"},"$02":{"mnemonic":"LD","length\
-":1,"cycles":[8],"flags":["-","-","-","-"],"addr":"$02","operand1":"(BC)","opera\
-nd2":"A"},"$03":{"mnemonic":"INC","length":1,"cycles":[8],"flags":["-","-","-","\
--"],"addr":"$03","operand1":"BC"},"$04":{"mnemonic":"INC","length":1,"cycles":[4\
-],"flags":["Z","0","H","-"],"addr":"$04","operand1":"B"},"$05":{"mnemonic":"DEC"\
-,"length":1,"cycles":[4],"flags":["Z","1","H","-"],"addr":"$05","operand1":"B"},\
-"$06":{"mnemonic":"LD","length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":\
-"$06","operand1":"B","operand2":"d8"},"$07":{"mnemonic":"RLCA","length":1,"cycle\
-s":[4],"flags":["0","0","0","C"],"addr":"$07"},"$08":{"mnemonic":"LD","length":3\
-,"cycles":[20],"flags":["-","-","-","-"],"addr":"$08","operand1":"(a16)","operan\
-d2":"SP"},"$09":{"mnemonic":"ADD","length":1,"cycles":[8],"flags":["-","0","H","\
-C"],"addr":"$09","operand1":"HL","operand2":"BC"},"$0a":{"mnemonic":"LD","length\
-":1,"cycles":[8],"flags":["-","-","-","-"],"addr":"$0a","operand1":"A","operand2\
-":"(BC)"},"$0b":{"mnemonic":"DEC","length":1,"cycles":[8],"flags":["-","-","-","\
--"],"addr":"$0b","operand1":"BC"},"$0c":{"mnemonic":"INC","length":1,"cycles":[4\
-],"flags":["Z","0","H","-"],"addr":"$0c","operand1":"C"},"$0d":{"mnemonic":"DEC"\
-,"length":1,"cycles":[4],"flags":["Z","1","H","-"],"addr":"$0d","operand1":"C"},\
-"$0e":{"mnemonic":"LD","length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":\
-"$0e","operand1":"C","operand2":"d8"},"$0f":{"mnemonic":"RRCA","length":1,"cycle\
-s":[4],"flags":["0","0","0","C"],"addr":"$0f"},"$10":{"mnemonic":"STOP","length"\
-:2,"cycles":[4],"flags":["-","-","-","-"],"addr":"$10","operand1":"0"},"$11":{"m\
-nemonic":"LD","length":3,"cycles":[12],"flags":["-","-","-","-"],"addr":"$11","o\
-perand1":"DE","operand2":"d16"},"$12":{"mnemonic":"LD","length":1,"cycles":[8],"\
-flags":["-","-","-","-"],"addr":"$12","operand1":"(DE)","operand2":"A"},"$13":{"\
-mnemonic":"INC","length":1,"cycles":[8],"flags":["-","-","-","-"],"addr":"$13","\
-operand1":"DE"},"$14":{"mnemonic":"INC","length":1,"cycles":[4],"flags":["Z","0"\
-,"H","-"],"addr":"$14","operand1":"D"},"$15":{"mnemonic":"DEC","length":1,"cycle\
-s":[4],"flags":["Z","1","H","-"],"addr":"$15","operand1":"D"},"$16":{"mnemonic":\
-"LD","length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$16","operand1":"\
-D","operand2":"d8"},"$17":{"mnemonic":"RLA","length":1,"cycles":[4],"flags":["0"\
-,"0","0","C"],"addr":"$17"},"$18":{"mnemonic":"JR","length":2,"cycles":[12],"fla\
-gs":["-","-","-","-"],"addr":"$18","operand1":"r8"},"$19":{"mnemonic":"ADD","len\
-gth":1,"cycles":[8],"flags":["-","0","H","C"],"addr":"$19","operand1":"HL","oper\
-and2":"DE"},"$1a":{"mnemonic":"LD","length":1,"cycles":[8],"flags":["-","-","-",\
-"-"],"addr":"$1a","operand1":"A","operand2":"(DE)"},"$1b":{"mnemonic":"DEC","len\
-gth":1,"cycles":[8],"flags":["-","-","-","-"],"addr":"$1b","operand1":"DE"},"$1c\
-":{"mnemonic":"INC","length":1,"cycles":[4],"flags":["Z","0","H","-"],"addr":"$1\
-c","operand1":"E"},"$1d":{"mnemonic":"DEC","length":1,"cycles":[4],"flags":["Z",\
-"1","H","-"],"addr":"$1d","operand1":"E"},"$1e":{"mnemonic":"LD","length":2,"cyc\
-les":[8],"flags":["-","-","-","-"],"addr":"$1e","operand1":"E","operand2":"d8"},\
-"$1f":{"mnemonic":"RRA","length":1,"cycles":[4],"flags":["0","0","0","C"],"addr"\
-:"$1f"},"$20":{"mnemonic":"JR","length":2,"cycles":[12,8],"flags":["-","-","-","\
--"],"addr":"$20","operand1":"NZ","operand2":"r8"},"$21":{"mnemonic":"LD","length\
-":3,"cycles":[12],"flags":["-","-","-","-"],"addr":"$21","operand1":"HL","operan\
-d2":"d16"},"$22":{"mnemonic":"LD","length":1,"cycles":[8],"flags":["-","-","-","\
--"],"addr":"$22","operand1":"(HL+)","operand2":"A"},"$23":{"mnemonic":"INC","len\
-gth":1,"cycles":[8],"flags":["-","-","-","-"],"addr":"$23","operand1":"HL"},"$24\
-":{"mnemonic":"INC","length":1,"cycles":[4],"flags":["Z","0","H","-"],"addr":"$2\
-4","operand1":"H"},"$25":{"mnemonic":"DEC","length":1,"cycles":[4],"flags":["Z",\
-"1","H","-"],"addr":"$25","operand1":"H"},"$26":{"mnemonic":"LD","length":2,"cyc\
-les":[8],"flags":["-","-","-","-"],"addr":"$26","operand1":"H","operand2":"d8"},\
-"$27":{"mnemonic":"DAA","length":1,"cycles":[4],"flags":["Z","-","0","C"],"addr"\
-:"$27"},"$28":{"mnemonic":"JR","length":2,"cycles":[12,8],"flags":["-","-","-","\
--"],"addr":"$28","operand1":"Z","operand2":"r8"},"$29":{"mnemonic":"ADD","length\
-":1,"cycles":[8],"flags":["-","0","H","C"],"addr":"$29","operand1":"HL","operand\
-2":"HL"},"$2a":{"mnemonic":"LD","length":1,"cycles":[8],"flags":["-","-","-","-"\
-],"addr":"$2a","operand1":"A","operand2":"(HL+)"},"$2b":{"mnemonic":"DEC","lengt\
-h":1,"cycles":[8],"flags":["-","-","-","-"],"addr":"$2b","operand1":"HL"},"$2c":\
-{"mnemonic":"INC","length":1,"cycles":[4],"flags":["Z","0","H","-"],"addr":"$2c"\
-,"operand1":"L"},"$2d":{"mnemonic":"DEC","length":1,"cycles":[4],"flags":["Z","1\
-","H","-"],"addr":"$2d","operand1":"L"},"$2e":{"mnemonic":"LD","length":2,"cycle\
-s":[8],"flags":["-","-","-","-"],"addr":"$2e","operand1":"L","operand2":"d8"},"$\
-2f":{"mnemonic":"CPL","length":1,"cycles":[4],"flags":["-","1","1","-"],"addr":"\
-$2f"},"$30":{"mnemonic":"JR","length":2,"cycles":[12,8],"flags":["-","-","-","-"\
-],"addr":"$30","operand1":"NC","operand2":"r8"},"$31":{"mnemonic":"LD","length":\
-3,"cycles":[12],"flags":["-","-","-","-"],"addr":"$31","operand1":"SP","operand2\
-":"d16"},"$32":{"mnemonic":"LD","length":1,"cycles":[8],"flags":["-","-","-","-"\
-],"addr":"$32","operand1":"(HL-)","operand2":"A"},"$33":{"mnemonic":"INC","lengt\
-h":1,"cycles":[8],"flags":["-","-","-","-"],"addr":"$33","operand1":"SP"},"$34":\
-{"mnemonic":"INC","length":1,"cycles":[12],"flags":["Z","0","H","-"],"addr":"$34\
-","operand1":"(HL)"},"$35":{"mnemonic":"DEC","length":1,"cycles":[12],"flags":["\
-Z","1","H","-"],"addr":"$35","operand1":"(HL)"},"$36":{"mnemonic":"LD","length":\
-2,"cycles":[12],"flags":["-","-","-","-"],"addr":"$36","operand1":"(HL)","operan\
-d2":"d8"},"$37":{"mnemonic":"SCF","length":1,"cycles":[4],"flags":["-","0","0","\
-1"],"addr":"$37"},"$38":{"mnemonic":"JR","length":2,"cycles":[12,8],"flags":["-"\
-,"-","-","-"],"addr":"$38","operand1":"C","operand2":"r8"},"$39":{"mnemonic":"AD\
-D","length":1,"cycles":[8],"flags":["-","0","H","C"],"addr":"$39","operand1":"HL\
-","operand2":"SP"},"$3a":{"mnemonic":"LD","length":1,"cycles":[8],"flags":["-","\
--","-","-"],"addr":"$3a","operand1":"A","operand2":"(HL-)"},"$3b":{"mnemonic":"D\
-EC","length":1,"cycles":[8],"flags":["-","-","-","-"],"addr":"$3b","operand1":"S\
-P"},"$3c":{"mnemonic":"INC","length":1,"cycles":[4],"flags":["Z","0","H","-"],"a\
-ddr":"$3c","operand1":"A"},"$3d":{"mnemonic":"DEC","length":1,"cycles":[4],"flag\
-s":["Z","1","H","-"],"addr":"$3d","operand1":"A"},"$3e":{"mnemonic":"LD","length\
-":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$3e","operand1":"A","operand2\
-":"d8"},"$3f":{"mnemonic":"CCF","length":1,"cycles":[4],"flags":["-","0","0","C"\
-],"addr":"$3f"},"$40":{"mnemonic":"LD","length":1,"cycles":[4],"flags":["-","-",\
-"-","-"],"addr":"$40","operand1":"B","operand2":"B"},"$41":{"mnemonic":"LD","len\
-gth":1,"cycles":[4],"flags":["-","-","-","-"],"addr":"$41","operand1":"B","opera\
-nd2":"C"},"$42":{"mnemonic":"LD","length":1,"cycles":[4],"flags":["-","-","-","-\
-"],"addr":"$42","operand1":"B","operand2":"D"},"$43":{"mnemonic":"LD","length":1\
-,"cycles":[4],"flags":["-","-","-","-"],"addr":"$43","operand1":"B","operand2":"\
-E"},"$44":{"mnemonic":"LD","length":1,"cycles":[4],"flags":["-","-","-","-"],"ad\
-dr":"$44","operand1":"B","operand2":"H"},"$45":{"mnemonic":"LD","length":1,"cycl\
-es":[4],"flags":["-","-","-","-"],"addr":"$45","operand1":"B","operand2":"L"},"$\
-46":{"mnemonic":"LD","length":1,"cycles":[8],"flags":["-","-","-","-"],"addr":"$\
-46","operand1":"B","operand2":"(HL)"},"$47":{"mnemonic":"LD","length":1,"cycles"\
-:[4],"flags":["-","-","-","-"],"addr":"$47","operand1":"B","operand2":"A"},"$48"\
-:{"mnemonic":"LD","length":1,"cycles":[4],"flags":["-","-","-","-"],"addr":"$48"\
-,"operand1":"C","operand2":"B"},"$49":{"mnemonic":"LD","length":1,"cycles":[4],"\
-flags":["-","-","-","-"],"addr":"$49","operand1":"C","operand2":"C"},"$4a":{"mne\
-monic":"LD","length":1,"cycles":[4],"flags":["-","-","-","-"],"addr":"$4a","oper\
-and1":"C","operand2":"D"},"$4b":{"mnemonic":"LD","length":1,"cycles":[4],"flags"\
-:["-","-","-","-"],"addr":"$4b","operand1":"C","operand2":"E"},"$4c":{"mnemonic"\
-:"LD","length":1,"cycles":[4],"flags":["-","-","-","-"],"addr":"$4c","operand1":\
-"C","operand2":"H"},"$4d":{"mnemonic":"LD","length":1,"cycles":[4],"flags":["-",\
-"-","-","-"],"addr":"$4d","operand1":"C","operand2":"L"},"$4e":{"mnemonic":"LD",\
-"length":1,"cycles":[8],"flags":["-","-","-","-"],"addr":"$4e","operand1":"C","o\
-perand2":"(HL)"},"$4f":{"mnemonic":"LD","length":1,"cycles":[4],"flags":["-","-"\
-,"-","-"],"addr":"$4f","operand1":"C","operand2":"A"},"$50":{"mnemonic":"LD","le\
-ngth":1,"cycles":[4],"flags":["-","-","-","-"],"addr":"$50","operand1":"D","oper\
-and2":"B"},"$51":{"mnemonic":"LD","length":1,"cycles":[4],"flags":["-","-","-","\
--"],"addr":"$51","operand1":"D","operand2":"C"},"$52":{"mnemonic":"LD","length":\
-1,"cycles":[4],"flags":["-","-","-","-"],"addr":"$52","operand1":"D","operand2":\
-"D"},"$53":{"mnemonic":"LD","length":1,"cycles":[4],"flags":["-","-","-","-"],"a\
-ddr":"$53","operand1":"D","operand2":"E"},"$54":{"mnemonic":"LD","length":1,"cyc\
-les":[4],"flags":["-","-","-","-"],"addr":"$54","operand1":"D","operand2":"H"},"\
-$55":{"mnemonic":"LD","length":1,"cycles":[4],"flags":["-","-","-","-"],"addr":"\
-$55","operand1":"D","operand2":"L"},"$56":{"mnemonic":"LD","length":1,"cycles":[\
-8],"flags":["-","-","-","-"],"addr":"$56","operand1":"D","operand2":"(HL)"},"$57\
-":{"mnemonic":"LD","length":1,"cycles":[4],"flags":["-","-","-","-"],"addr":"$57\
-","operand1":"D","operand2":"A"},"$58":{"mnemonic":"LD","length":1,"cycles":[4],\
-"flags":["-","-","-","-"],"addr":"$58","operand1":"E","operand2":"B"},"$59":{"mn\
-emonic":"LD","length":1,"cycles":[4],"flags":["-","-","-","-"],"addr":"$59","ope\
-rand1":"E","operand2":"C"},"$5a":{"mnemonic":"LD","length":1,"cycles":[4],"flags\
-":["-","-","-","-"],"addr":"$5a","operand1":"E","operand2":"D"},"$5b":{"mnemonic\
-":"LD","length":1,"cycles":[4],"flags":["-","-","-","-"],"addr":"$5b","operand1"\
-:"E","operand2":"E"},"$5c":{"mnemonic":"LD","length":1,"cycles":[4],"flags":["-"\
-,"-","-","-"],"addr":"$5c","operand1":"E","operand2":"H"},"$5d":{"mnemonic":"LD"\
-,"length":1,"cycles":[4],"flags":["-","-","-","-"],"addr":"$5d","operand1":"E","\
-operand2":"L"},"$5e":{"mnemonic":"LD","length":1,"cycles":[8],"flags":["-","-","\
--","-"],"addr":"$5e","operand1":"E","operand2":"(HL)"},"$5f":{"mnemonic":"LD","l\
-ength":1,"cycles":[4],"flags":["-","-","-","-"],"addr":"$5f","operand1":"E","ope\
-rand2":"A"},"$60":{"mnemonic":"LD","length":1,"cycles":[4],"flags":["-","-","-",\
-"-"],"addr":"$60","operand1":"H","operand2":"B"},"$61":{"mnemonic":"LD","length"\
-:1,"cycles":[4],"flags":["-","-","-","-"],"addr":"$61","operand1":"H","operand2"\
-:"C"},"$62":{"mnemonic":"LD","length":1,"cycles":[4],"flags":["-","-","-","-"],"\
-addr":"$62","operand1":"H","operand2":"D"},"$63":{"mnemonic":"LD","length":1,"cy\
-cles":[4],"flags":["-","-","-","-"],"addr":"$63","operand1":"H","operand2":"E"},\
-"$64":{"mnemonic":"LD","length":1,"cycles":[4],"flags":["-","-","-","-"],"addr":\
-"$64","operand1":"H","operand2":"H"},"$65":{"mnemonic":"LD","length":1,"cycles":\
-[4],"flags":["-","-","-","-"],"addr":"$65","operand1":"H","operand2":"L"},"$66":\
-{"mnemonic":"LD","length":1,"cycles":[8],"flags":["-","-","-","-"],"addr":"$66",\
-"operand1":"H","operand2":"(HL)"},"$67":{"mnemonic":"LD","length":1,"cycles":[4]\
-,"flags":["-","-","-","-"],"addr":"$67","operand1":"H","operand2":"A"},"$68":{"m\
-nemonic":"LD","length":1,"cycles":[4],"flags":["-","-","-","-"],"addr":"$68","op\
-erand1":"L","operand2":"B"},"$69":{"mnemonic":"LD","length":1,"cycles":[4],"flag\
-s":["-","-","-","-"],"addr":"$69","operand1":"L","operand2":"C"},"$6a":{"mnemoni\
-c":"LD","length":1,"cycles":[4],"flags":["-","-","-","-"],"addr":"$6a","operand1\
-":"L","operand2":"D"},"$6b":{"mnemonic":"LD","length":1,"cycles":[4],"flags":["-\
-","-","-","-"],"addr":"$6b","operand1":"L","operand2":"E"},"$6c":{"mnemonic":"LD\
-","length":1,"cycles":[4],"flags":["-","-","-","-"],"addr":"$6c","operand1":"L",\
-"operand2":"H"},"$6d":{"mnemonic":"LD","length":1,"cycles":[4],"flags":["-","-",\
-"-","-"],"addr":"$6d","operand1":"L","operand2":"L"},"$6e":{"mnemonic":"LD","len\
-gth":1,"cycles":[8],"flags":["-","-","-","-"],"addr":"$6e","operand1":"L","opera\
-nd2":"(HL)"},"$6f":{"mnemonic":"LD","length":1,"cycles":[4],"flags":["-","-","-"\
-,"-"],"addr":"$6f","operand1":"L","operand2":"A"},"$70":{"mnemonic":"LD","length\
-":1,"cycles":[8],"flags":["-","-","-","-"],"addr":"$70","operand1":"(HL)","opera\
-nd2":"B"},"$71":{"mnemonic":"LD","length":1,"cycles":[8],"flags":["-","-","-","-\
-"],"addr":"$71","operand1":"(HL)","operand2":"C"},"$72":{"mnemonic":"LD","length\
-":1,"cycles":[8],"flags":["-","-","-","-"],"addr":"$72","operand1":"(HL)","opera\
-nd2":"D"},"$73":{"mnemonic":"LD","length":1,"cycles":[8],"flags":["-","-","-","-\
-"],"addr":"$73","operand1":"(HL)","operand2":"E"},"$74":{"mnemonic":"LD","length\
-":1,"cycles":[8],"flags":["-","-","-","-"],"addr":"$74","operand1":"(HL)","opera\
-nd2":"H"},"$75":{"mnemonic":"LD","length":1,"cycles":[8],"flags":["-","-","-","-\
-"],"addr":"$75","operand1":"(HL)","operand2":"L"},"$76":{"mnemonic":"HALT","leng\
-th":1,"cycles":[4],"flags":["-","-","-","-"],"addr":"$76"},"$77":{"mnemonic":"LD\
-","length":1,"cycles":[8],"flags":["-","-","-","-"],"addr":"$77","operand1":"(HL\
-)","operand2":"A"},"$78":{"mnemonic":"LD","length":1,"cycles":[4],"flags":["-","\
--","-","-"],"addr":"$78","operand1":"A","operand2":"B"},"$79":{"mnemonic":"LD","\
-length":1,"cycles":[4],"flags":["-","-","-","-"],"addr":"$79","operand1":"A","op\
-erand2":"C"},"$7a":{"mnemonic":"LD","length":1,"cycles":[4],"flags":["-","-","-"\
-,"-"],"addr":"$7a","operand1":"A","operand2":"D"},"$7b":{"mnemonic":"LD","length\
-":1,"cycles":[4],"flags":["-","-","-","-"],"addr":"$7b","operand1":"A","operand2\
-":"E"},"$7c":{"mnemonic":"LD","length":1,"cycles":[4],"flags":["-","-","-","-"],\
-"addr":"$7c","operand1":"A","operand2":"H"},"$7d":{"mnemonic":"LD","length":1,"c\
-ycles":[4],"flags":["-","-","-","-"],"addr":"$7d","operand1":"A","operand2":"L"}\
-,"$7e":{"mnemonic":"LD","length":1,"cycles":[8],"flags":["-","-","-","-"],"addr"\
-:"$7e","operand1":"A","operand2":"(HL)"},"$7f":{"mnemonic":"LD","length":1,"cycl\
-es":[4],"flags":["-","-","-","-"],"addr":"$7f","operand1":"A","operand2":"A"},"$\
-80":{"mnemonic":"ADD","length":1,"cycles":[4],"flags":["Z","0","H","C"],"addr":"\
-$80","operand1":"A","operand2":"B"},"$81":{"mnemonic":"ADD","length":1,"cycles":\
-[4],"flags":["Z","0","H","C"],"addr":"$81","operand1":"A","operand2":"C"},"$82":\
-{"mnemonic":"ADD","length":1,"cycles":[4],"flags":["Z","0","H","C"],"addr":"$82"\
-,"operand1":"A","operand2":"D"},"$83":{"mnemonic":"ADD","length":1,"cycles":[4],\
-"flags":["Z","0","H","C"],"addr":"$83","operand1":"A","operand2":"E"},"$84":{"mn\
-emonic":"ADD","length":1,"cycles":[4],"flags":["Z","0","H","C"],"addr":"$84","op\
-erand1":"A","operand2":"H"},"$85":{"mnemonic":"ADD","length":1,"cycles":[4],"fla\
-gs":["Z","0","H","C"],"addr":"$85","operand1":"A","operand2":"L"},"$86":{"mnemon\
-ic":"ADD","length":1,"cycles":[8],"flags":["Z","0","H","C"],"addr":"$86","operan\
-d1":"A","operand2":"(HL)"},"$87":{"mnemonic":"ADD","length":1,"cycles":[4],"flag\
-s":["Z","0","H","C"],"addr":"$87","operand1":"A","operand2":"A"},"$88":{"mnemoni\
-c":"ADC","length":1,"cycles":[4],"flags":["Z","0","H","C"],"addr":"$88","operand\
-1":"A","operand2":"B"},"$89":{"mnemonic":"ADC","length":1,"cycles":[4],"flags":[\
-"Z","0","H","C"],"addr":"$89","operand1":"A","operand2":"C"},"$8a":{"mnemonic":"\
-ADC","length":1,"cycles":[4],"flags":["Z","0","H","C"],"addr":"$8a","operand1":"\
-A","operand2":"D"},"$8b":{"mnemonic":"ADC","length":1,"cycles":[4],"flags":["Z",\
-"0","H","C"],"addr":"$8b","operand1":"A","operand2":"E"},"$8c":{"mnemonic":"ADC"\
-,"length":1,"cycles":[4],"flags":["Z","0","H","C"],"addr":"$8c","operand1":"A","\
-operand2":"H"},"$8d":{"mnemonic":"ADC","length":1,"cycles":[4],"flags":["Z","0",\
-"H","C"],"addr":"$8d","operand1":"A","operand2":"L"},"$8e":{"mnemonic":"ADC","le\
-ngth":1,"cycles":[8],"flags":["Z","0","H","C"],"addr":"$8e","operand1":"A","oper\
-and2":"(HL)"},"$8f":{"mnemonic":"ADC","length":1,"cycles":[4],"flags":["Z","0","\
-H","C"],"addr":"$8f","operand1":"A","operand2":"A"},"$90":{"mnemonic":"SUB","len\
-gth":1,"cycles":[4],"flags":["Z","1","H","C"],"addr":"$90","operand1":"B"},"$91"\
-:{"mnemonic":"SUB","length":1,"cycles":[4],"flags":["Z","1","H","C"],"addr":"$91\
-","operand1":"C"},"$92":{"mnemonic":"SUB","length":1,"cycles":[4],"flags":["Z","\
-1","H","C"],"addr":"$92","operand1":"D"},"$93":{"mnemonic":"SUB","length":1,"cyc\
-les":[4],"flags":["Z","1","H","C"],"addr":"$93","operand1":"E"},"$94":{"mnemonic\
-":"SUB","length":1,"cycles":[4],"flags":["Z","1","H","C"],"addr":"$94","operand1\
-":"H"},"$95":{"mnemonic":"SUB","length":1,"cycles":[4],"flags":["Z","1","H","C"]\
-,"addr":"$95","operand1":"L"},"$96":{"mnemonic":"SUB","length":1,"cycles":[8],"f\
-lags":["Z","1","H","C"],"addr":"$96","operand1":"(HL)"},"$97":{"mnemonic":"SUB",\
-"length":1,"cycles":[4],"flags":["Z","1","H","C"],"addr":"$97","operand1":"A"},"\
-$98":{"mnemonic":"SBC","length":1,"cycles":[4],"flags":["Z","1","H","C"],"addr":\
-"$98","operand1":"A","operand2":"B"},"$99":{"mnemonic":"SBC","length":1,"cycles"\
-:[4],"flags":["Z","1","H","C"],"addr":"$99","operand1":"A","operand2":"C"},"$9a"\
-:{"mnemonic":"SBC","length":1,"cycles":[4],"flags":["Z","1","H","C"],"addr":"$9a\
-","operand1":"A","operand2":"D"},"$9b":{"mnemonic":"SBC","length":1,"cycles":[4]\
-,"flags":["Z","1","H","C"],"addr":"$9b","operand1":"A","operand2":"E"},"$9c":{"m\
-nemonic":"SBC","length":1,"cycles":[4],"flags":["Z","1","H","C"],"addr":"$9c","o\
-perand1":"A","operand2":"H"},"$9d":{"mnemonic":"SBC","length":1,"cycles":[4],"fl\
-ags":["Z","1","H","C"],"addr":"$9d","operand1":"A","operand2":"L"},"$9e":{"mnemo\
-nic":"SBC","length":1,"cycles":[8],"flags":["Z","1","H","C"],"addr":"$9e","opera\
-nd1":"A","operand2":"(HL)"},"$9f":{"mnemonic":"SBC","length":1,"cycles":[4],"fla\
-gs":["Z","1","H","C"],"addr":"$9f","operand1":"A","operand2":"A"},"$a0":{"mnemon\
-ic":"AND","length":1,"cycles":[4],"flags":["Z","0","1","0"],"addr":"$a0","operan\
-d1":"B"},"$a1":{"mnemonic":"AND","length":1,"cycles":[4],"flags":["Z","0","1","0\
-"],"addr":"$a1","operand1":"C"},"$a2":{"mnemonic":"AND","length":1,"cycles":[4],\
-"flags":["Z","0","1","0"],"addr":"$a2","operand1":"D"},"$a3":{"mnemonic":"AND","\
-length":1,"cycles":[4],"flags":["Z","0","1","0"],"addr":"$a3","operand1":"E"},"$\
-a4":{"mnemonic":"AND","length":1,"cycles":[4],"flags":["Z","0","1","0"],"addr":"\
-$a4","operand1":"H"},"$a5":{"mnemonic":"AND","length":1,"cycles":[4],"flags":["Z\
-","0","1","0"],"addr":"$a5","operand1":"L"},"$a6":{"mnemonic":"AND","length":1,"\
-cycles":[8],"flags":["Z","0","1","0"],"addr":"$a6","operand1":"(HL)"},"$a7":{"mn\
-emonic":"AND","length":1,"cycles":[4],"flags":["Z","0","1","0"],"addr":"$a7","op\
-erand1":"A"},"$a8":{"mnemonic":"XOR","length":1,"cycles":[4],"flags":["Z","0","0\
-","0"],"addr":"$a8","operand1":"B"},"$a9":{"mnemonic":"XOR","length":1,"cycles":\
-[4],"flags":["Z","0","0","0"],"addr":"$a9","operand1":"C"},"$aa":{"mnemonic":"XO\
-R","length":1,"cycles":[4],"flags":["Z","0","0","0"],"addr":"$aa","operand1":"D"\
-},"$ab":{"mnemonic":"XOR","length":1,"cycles":[4],"flags":["Z","0","0","0"],"add\
-r":"$ab","operand1":"E"},"$ac":{"mnemonic":"XOR","length":1,"cycles":[4],"flags"\
-:["Z","0","0","0"],"addr":"$ac","operand1":"H"},"$ad":{"mnemonic":"XOR","length"\
-:1,"cycles":[4],"flags":["Z","0","0","0"],"addr":"$ad","operand1":"L"},"$ae":{"m\
-nemonic":"XOR","length":1,"cycles":[8],"flags":["Z","0","0","0"],"addr":"$ae","o\
-perand1":"(HL)"},"$af":{"mnemonic":"XOR","length":1,"cycles":[4],"flags":["Z","0\
-","0","0"],"addr":"$af","operand1":"A"},"$b0":{"mnemonic":"OR","length":1,"cycle\
-s":[4],"flags":["Z","0","0","0"],"addr":"$b0","operand1":"B"},"$b1":{"mnemonic":\
-"OR","length":1,"cycles":[4],"flags":["Z","0","0","0"],"addr":"$b1","operand1":"\
-C"},"$b2":{"mnemonic":"OR","length":1,"cycles":[4],"flags":["Z","0","0","0"],"ad\
-dr":"$b2","operand1":"D"},"$b3":{"mnemonic":"OR","length":1,"cycles":[4],"flags"\
-:["Z","0","0","0"],"addr":"$b3","operand1":"E"},"$b4":{"mnemonic":"OR","length":\
-1,"cycles":[4],"flags":["Z","0","0","0"],"addr":"$b4","operand1":"H"},"$b5":{"mn\
-emonic":"OR","length":1,"cycles":[4],"flags":["Z","0","0","0"],"addr":"$b5","ope\
-rand1":"L"},"$b6":{"mnemonic":"OR","length":1,"cycles":[8],"flags":["Z","0","0",\
-"0"],"addr":"$b6","operand1":"(HL)"},"$b7":{"mnemonic":"OR","length":1,"cycles":\
-[4],"flags":["Z","0","0","0"],"addr":"$b7","operand1":"A"},"$b8":{"mnemonic":"CP\
-","length":1,"cycles":[4],"flags":["Z","1","H","C"],"addr":"$b8","operand1":"B"}\
-,"$b9":{"mnemonic":"CP","length":1,"cycles":[4],"flags":["Z","1","H","C"],"addr"\
-:"$b9","operand1":"C"},"$ba":{"mnemonic":"CP","length":1,"cycles":[4],"flags":["\
-Z","1","H","C"],"addr":"$ba","operand1":"D"},"$bb":{"mnemonic":"CP","length":1,"\
-cycles":[4],"flags":["Z","1","H","C"],"addr":"$bb","operand1":"E"},"$bc":{"mnemo\
-nic":"CP","length":1,"cycles":[4],"flags":["Z","1","H","C"],"addr":"$bc","operan\
-d1":"H"},"$bd":{"mnemonic":"CP","length":1,"cycles":[4],"flags":["Z","1","H","C"\
-],"addr":"$bd","operand1":"L"},"$be":{"mnemonic":"CP","length":1,"cycles":[8],"f\
-lags":["Z","1","H","C"],"addr":"$be","operand1":"(HL)"},"$bf":{"mnemonic":"CP","\
-length":1,"cycles":[4],"flags":["Z","1","H","C"],"addr":"$bf","operand1":"A"},"$\
-c0":{"mnemonic":"RET","length":1,"cycles":[20,8],"flags":["-","-","-","-"],"addr\
-":"$c0","operand1":"NZ"},"$c1":{"mnemonic":"POP","length":1,"cycles":[12],"flags\
-":["-","-","-","-"],"addr":"$c1","operand1":"BC"},"$c2":{"mnemonic":"JP","length\
-":3,"cycles":[16,12],"flags":["-","-","-","-"],"addr":"$c2","operand1":"NZ","ope\
-rand2":"a16"},"$c3":{"mnemonic":"JP","length":3,"cycles":[16],"flags":["-","-","\
--","-"],"addr":"$c3","operand1":"a16"},"$c4":{"mnemonic":"CALL","length":3,"cycl\
-es":[24,12],"flags":["-","-","-","-"],"addr":"$c4","operand1":"NZ","operand2":"a\
-16"},"$c5":{"mnemonic":"PUSH","length":1,"cycles":[16],"flags":["-","-","-","-"]\
-,"addr":"$c5","operand1":"BC"},"$c6":{"mnemonic":"ADD","length":2,"cycles":[8],"\
-flags":["Z","0","H","C"],"addr":"$c6","operand1":"A","operand2":"d8"},"$c7":{"mn\
-emonic":"RST","length":1,"cycles":[16],"flags":["-","-","-","-"],"addr":"$c7","o\
-perand1":"00H"},"$c8":{"mnemonic":"RET","length":1,"cycles":[20,8],"flags":["-",\
-"-","-","-"],"addr":"$c8","operand1":"Z"},"$c9":{"mnemonic":"RET","length":1,"cy\
-cles":[16],"flags":["-","-","-","-"],"addr":"$c9"},"$ca":{"mnemonic":"JP","lengt\
-h":3,"cycles":[16,12],"flags":["-","-","-","-"],"addr":"$ca","operand1":"Z","ope\
-rand2":"a16"},"$cb":{"mnemonic":"PREFIX","length":1,"cycles":[4],"flags":["-","-\
-","-","-"],"addr":"$cb","operand1":"CB"},"$cc":{"mnemonic":"CALL","length":3,"cy\
-cles":[24,12],"flags":["-","-","-","-"],"addr":"$cc","operand1":"Z","operand2":"\
-a16"},"$cd":{"mnemonic":"CALL","length":3,"cycles":[24],"flags":["-","-","-","-"\
-],"addr":"$cd","operand1":"a16"},"$ce":{"mnemonic":"ADC","length":2,"cycles":[8]\
-,"flags":["Z","0","H","C"],"addr":"$ce","operand1":"A","operand2":"d8"},"$cf":{"\
-mnemonic":"RST","length":1,"cycles":[16],"flags":["-","-","-","-"],"addr":"$cf",\
-"operand1":"08H"},"$d0":{"mnemonic":"RET","length":1,"cycles":[20,8],"flags":["-\
-","-","-","-"],"addr":"$d0","operand1":"NC"},"$d1":{"mnemonic":"POP","length":1,\
-"cycles":[12],"flags":["-","-","-","-"],"addr":"$d1","operand1":"DE"},"$d2":{"mn\
-emonic":"JP","length":3,"cycles":[16,12],"flags":["-","-","-","-"],"addr":"$d2",\
-"operand1":"NC","operand2":"a16"},"$d4":{"mnemonic":"CALL","length":3,"cycles":[\
-24,12],"flags":["-","-","-","-"],"addr":"$d4","operand1":"NC","operand2":"a16"},\
-"$d5":{"mnemonic":"PUSH","length":1,"cycles":[16],"flags":["-","-","-","-"],"add\
-r":"$d5","operand1":"DE"},"$d6":{"mnemonic":"SUB","length":2,"cycles":[8],"flags\
-":["Z","1","H","C"],"addr":"$d6","operand1":"d8"},"$d7":{"mnemonic":"RST","lengt\
-h":1,"cycles":[16],"flags":["-","-","-","-"],"addr":"$d7","operand1":"10H"},"$d8\
-":{"mnemonic":"RET","length":1,"cycles":[20,8],"flags":["-","-","-","-"],"addr":\
-"$d8","operand1":"C"},"$d9":{"mnemonic":"RETI","length":1,"cycles":[16],"flags":\
-["-","-","-","-"],"addr":"$d9"},"$da":{"mnemonic":"JP","length":3,"cycles":[16,1\
-2],"flags":["-","-","-","-"],"addr":"$da","operand1":"C","operand2":"a16"},"$dc"\
-:{"mnemonic":"CALL","length":3,"cycles":[24,12],"flags":["-","-","-","-"],"addr"\
-:"$dc","operand1":"C","operand2":"a16"},"$de":{"mnemonic":"SBC","length":2,"cycl\
-es":[8],"flags":["Z","1","H","C"],"addr":"$de","operand1":"A","operand2":"d8"},"\
-$df":{"mnemonic":"RST","length":1,"cycles":[16],"flags":["-","-","-","-"],"addr"\
-:"$df","operand1":"18H"},"$e0":{"mnemonic":"LDH","length":2,"cycles":[12],"flags\
-":["-","-","-","-"],"addr":"$e0","operand1":"(a8)","operand2":"A"},"$e1":{"mnemo\
-nic":"POP","length":1,"cycles":[12],"flags":["-","-","-","-"],"addr":"$e1","oper\
-and1":"HL"},"$e2":{"mnemonic":"LD","length":2,"cycles":[8],"flags":["-","-","-",\
-"-"],"addr":"$e2","operand1":"(C)","operand2":"A"},"$e5":{"mnemonic":"PUSH","len\
-gth":1,"cycles":[16],"flags":["-","-","-","-"],"addr":"$e5","operand1":"HL"},"$e\
-6":{"mnemonic":"AND","length":2,"cycles":[8],"flags":["Z","0","1","0"],"addr":"$\
-e6","operand1":"d8"},"$e7":{"mnemonic":"RST","length":1,"cycles":[16],"flags":["\
--","-","-","-"],"addr":"$e7","operand1":"20H"},"$e8":{"mnemonic":"ADD","length":\
-2,"cycles":[16],"flags":["0","0","H","C"],"addr":"$e8","operand1":"SP","operand2\
-":"r8"},"$e9":{"mnemonic":"JP","length":1,"cycles":[4],"flags":["-","-","-","-"]\
-,"addr":"$e9","operand1":"(HL)"},"$ea":{"mnemonic":"LD","length":3,"cycles":[16]\
-,"flags":["-","-","-","-"],"addr":"$ea","operand1":"(a16)","operand2":"A"},"$ee"\
-:{"mnemonic":"XOR","length":2,"cycles":[8],"flags":["Z","0","0","0"],"addr":"$ee\
-","operand1":"d8"},"$ef":{"mnemonic":"RST","length":1,"cycles":[16],"flags":["-"\
-,"-","-","-"],"addr":"$ef","operand1":"28H"},"$f0":{"mnemonic":"LDH","length":2,\
-"cycles":[12],"flags":["-","-","-","-"],"addr":"$f0","operand1":"A","operand2":"\
-(a8)"},"$f1":{"mnemonic":"POP","length":1,"cycles":[12],"flags":["Z","N","H","C"\
-],"addr":"$f1","operand1":"AF"},"$f2":{"mnemonic":"LD","length":2,"cycles":[8],"\
-flags":["-","-","-","-"],"addr":"$f2","operand1":"A","operand2":"(C)"},"$f3":{"m\
-nemonic":"DI","length":1,"cycles":[4],"flags":["-","-","-","-"],"addr":"$f3"},"$\
-f5":{"mnemonic":"PUSH","length":1,"cycles":[16],"flags":["-","-","-","-"],"addr"\
-:"$f5","operand1":"AF"},"$f6":{"mnemonic":"OR","length":2,"cycles":[8],"flags":[\
-"Z","0","0","0"],"addr":"$f6","operand1":"d8"},"$f7":{"mnemonic":"RST","length":\
-1,"cycles":[16],"flags":["-","-","-","-"],"addr":"$f7","operand1":"30H"},"$f8":{\
-"mnemonic":"LD","length":2,"cycles":[12],"flags":["0","0","H","C"],"addr":"$f8",\
-"operand1":"HL","operand2":"SP+r8"},"$f9":{"mnemonic":"LD","length":1,"cycles":[\
-8],"flags":["-","-","-","-"],"addr":"$f9","operand1":"SP","operand2":"HL"},"$fa"\
-:{"mnemonic":"LD","length":3,"cycles":[16],"flags":["-","-","-","-"],"addr":"$fa\
-","operand1":"A","operand2":"(a16)"},"$fb":{"mnemonic":"EI","length":1,"cycles":\
-[4],"flags":["-","-","-","-"],"addr":"$fb"},"$fe":{"mnemonic":"CP","length":2,"c\
-ycles":[8],"flags":["Z","1","H","C"],"addr":"$fe","operand1":"d8"},"$ff":{"mnemo\
-nic":"RST","length":1,"cycles":[16],"flags":["-","-","-","-"],"addr":"$ff","oper\
-and1":"38H"},"$cb00":{"mnemonic":"RLC","length":2,"cycles":[8],"flags":["Z","0",\
-"0","C"],"addr":"$cb00","operand1":"B"},"$cb01":{"mnemonic":"RLC","length":2,"cy\
-cles":[8],"flags":["Z","0","0","C"],"addr":"$cb01","operand1":"C"},"$cb02":{"mne\
-monic":"RLC","length":2,"cycles":[8],"flags":["Z","0","0","C"],"addr":"$cb02","o\
-perand1":"D"},"$cb03":{"mnemonic":"RLC","length":2,"cycles":[8],"flags":["Z","0"\
-,"0","C"],"addr":"$cb03","operand1":"E"},"$cb04":{"mnemonic":"RLC","length":2,"c\
-ycles":[8],"flags":["Z","0","0","C"],"addr":"$cb04","operand1":"H"},"$cb05":{"mn\
-emonic":"RLC","length":2,"cycles":[8],"flags":["Z","0","0","C"],"addr":"$cb05","\
-operand1":"L"},"$cb06":{"mnemonic":"RLC","length":2,"cycles":[16],"flags":["Z","\
-0","0","C"],"addr":"$cb06","operand1":"(HL)"},"$cb07":{"mnemonic":"RLC","length"\
-:2,"cycles":[8],"flags":["Z","0","0","C"],"addr":"$cb07","operand1":"A"},"$cb08"\
-:{"mnemonic":"RRC","length":2,"cycles":[8],"flags":["Z","0","0","C"],"addr":"$cb\
-08","operand1":"B"},"$cb09":{"mnemonic":"RRC","length":2,"cycles":[8],"flags":["\
-Z","0","0","C"],"addr":"$cb09","operand1":"C"},"$cb0a":{"mnemonic":"RRC","length\
-":2,"cycles":[8],"flags":["Z","0","0","C"],"addr":"$cb0a","operand1":"D"},"$cb0b\
-":{"mnemonic":"RRC","length":2,"cycles":[8],"flags":["Z","0","0","C"],"addr":"$c\
-b0b","operand1":"E"},"$cb0c":{"mnemonic":"RRC","length":2,"cycles":[8],"flags":[\
-"Z","0","0","C"],"addr":"$cb0c","operand1":"H"},"$cb0d":{"mnemonic":"RRC","lengt\
-h":2,"cycles":[8],"flags":["Z","0","0","C"],"addr":"$cb0d","operand1":"L"},"$cb0\
-e":{"mnemonic":"RRC","length":2,"cycles":[16],"flags":["Z","0","0","C"],"addr":"\
-$cb0e","operand1":"(HL)"},"$cb0f":{"mnemonic":"RRC","length":2,"cycles":[8],"fla\
-gs":["Z","0","0","C"],"addr":"$cb0f","operand1":"A"},"$cb10":{"mnemonic":"RL","l\
-ength":2,"cycles":[8],"flags":["Z","0","0","C"],"addr":"$cb10","operand1":"B"},"\
-$cb11":{"mnemonic":"RL","length":2,"cycles":[8],"flags":["Z","0","0","C"],"addr"\
-:"$cb11","operand1":"C"},"$cb12":{"mnemonic":"RL","length":2,"cycles":[8],"flags\
-":["Z","0","0","C"],"addr":"$cb12","operand1":"D"},"$cb13":{"mnemonic":"RL","len\
-gth":2,"cycles":[8],"flags":["Z","0","0","C"],"addr":"$cb13","operand1":"E"},"$c\
-b14":{"mnemonic":"RL","length":2,"cycles":[8],"flags":["Z","0","0","C"],"addr":"\
-$cb14","operand1":"H"},"$cb15":{"mnemonic":"RL","length":2,"cycles":[8],"flags":\
-["Z","0","0","C"],"addr":"$cb15","operand1":"L"},"$cb16":{"mnemonic":"RL","lengt\
-h":2,"cycles":[16],"flags":["Z","0","0","C"],"addr":"$cb16","operand1":"(HL)"},"\
-$cb17":{"mnemonic":"RL","length":2,"cycles":[8],"flags":["Z","0","0","C"],"addr"\
-:"$cb17","operand1":"A"},"$cb18":{"mnemonic":"RR","length":2,"cycles":[8],"flags\
-":["Z","0","0","C"],"addr":"$cb18","operand1":"B"},"$cb19":{"mnemonic":"RR","len\
-gth":2,"cycles":[8],"flags":["Z","0","0","C"],"addr":"$cb19","operand1":"C"},"$c\
-b1a":{"mnemonic":"RR","length":2,"cycles":[8],"flags":["Z","0","0","C"],"addr":"\
-$cb1a","operand1":"D"},"$cb1b":{"mnemonic":"RR","length":2,"cycles":[8],"flags":\
-["Z","0","0","C"],"addr":"$cb1b","operand1":"E"},"$cb1c":{"mnemonic":"RR","lengt\
-h":2,"cycles":[8],"flags":["Z","0","0","C"],"addr":"$cb1c","operand1":"H"},"$cb1\
-d":{"mnemonic":"RR","length":2,"cycles":[8],"flags":["Z","0","0","C"],"addr":"$c\
-b1d","operand1":"L"},"$cb1e":{"mnemonic":"RR","length":2,"cycles":[16],"flags":[\
-"Z","0","0","C"],"addr":"$cb1e","operand1":"(HL)"},"$cb1f":{"mnemonic":"RR","len\
-gth":2,"cycles":[8],"flags":["Z","0","0","C"],"addr":"$cb1f","operand1":"A"},"$c\
-b20":{"mnemonic":"SLA","length":2,"cycles":[8],"flags":["Z","0","0","C"],"addr":\
-"$cb20","operand1":"B"},"$cb21":{"mnemonic":"SLA","length":2,"cycles":[8],"flags\
-":["Z","0","0","C"],"addr":"$cb21","operand1":"C"},"$cb22":{"mnemonic":"SLA","le\
-ngth":2,"cycles":[8],"flags":["Z","0","0","C"],"addr":"$cb22","operand1":"D"},"$\
-cb23":{"mnemonic":"SLA","length":2,"cycles":[8],"flags":["Z","0","0","C"],"addr"\
-:"$cb23","operand1":"E"},"$cb24":{"mnemonic":"SLA","length":2,"cycles":[8],"flag\
-s":["Z","0","0","C"],"addr":"$cb24","operand1":"H"},"$cb25":{"mnemonic":"SLA","l\
-ength":2,"cycles":[8],"flags":["Z","0","0","C"],"addr":"$cb25","operand1":"L"},"\
-$cb26":{"mnemonic":"SLA","length":2,"cycles":[16],"flags":["Z","0","0","C"],"add\
-r":"$cb26","operand1":"(HL)"},"$cb27":{"mnemonic":"SLA","length":2,"cycles":[8],\
-"flags":["Z","0","0","C"],"addr":"$cb27","operand1":"A"},"$cb28":{"mnemonic":"SR\
-A","length":2,"cycles":[8],"flags":["Z","0","0","0"],"addr":"$cb28","operand1":"\
-B"},"$cb29":{"mnemonic":"SRA","length":2,"cycles":[8],"flags":["Z","0","0","0"],\
-"addr":"$cb29","operand1":"C"},"$cb2a":{"mnemonic":"SRA","length":2,"cycles":[8]\
-,"flags":["Z","0","0","0"],"addr":"$cb2a","operand1":"D"},"$cb2b":{"mnemonic":"S\
-RA","length":2,"cycles":[8],"flags":["Z","0","0","0"],"addr":"$cb2b","operand1":\
-"E"},"$cb2c":{"mnemonic":"SRA","length":2,"cycles":[8],"flags":["Z","0","0","0"]\
-,"addr":"$cb2c","operand1":"H"},"$cb2d":{"mnemonic":"SRA","length":2,"cycles":[8\
-],"flags":["Z","0","0","0"],"addr":"$cb2d","operand1":"L"},"$cb2e":{"mnemonic":"\
-SRA","length":2,"cycles":[16],"flags":["Z","0","0","0"],"addr":"$cb2e","operand1\
-":"(HL)"},"$cb2f":{"mnemonic":"SRA","length":2,"cycles":[8],"flags":["Z","0","0"\
-,"0"],"addr":"$cb2f","operand1":"A"},"$cb30":{"mnemonic":"SWAP","length":2,"cycl\
-es":[8],"flags":["Z","0","0","0"],"addr":"$cb30","operand1":"B"},"$cb31":{"mnemo\
-nic":"SWAP","length":2,"cycles":[8],"flags":["Z","0","0","0"],"addr":"$cb31","op\
-erand1":"C"},"$cb32":{"mnemonic":"SWAP","length":2,"cycles":[8],"flags":["Z","0"\
-,"0","0"],"addr":"$cb32","operand1":"D"},"$cb33":{"mnemonic":"SWAP","length":2,"\
-cycles":[8],"flags":["Z","0","0","0"],"addr":"$cb33","operand1":"E"},"$cb34":{"m\
-nemonic":"SWAP","length":2,"cycles":[8],"flags":["Z","0","0","0"],"addr":"$cb34"\
-,"operand1":"H"},"$cb35":{"mnemonic":"SWAP","length":2,"cycles":[8],"flags":["Z"\
-,"0","0","0"],"addr":"$cb35","operand1":"L"},"$cb36":{"mnemonic":"SWAP","length"\
-:2,"cycles":[16],"flags":["Z","0","0","0"],"addr":"$cb36","operand1":"(HL)"},"$c\
-b37":{"mnemonic":"SWAP","length":2,"cycles":[8],"flags":["Z","0","0","0"],"addr"\
-:"$cb37","operand1":"A"},"$cb38":{"mnemonic":"SRL","length":2,"cycles":[8],"flag\
-s":["Z","0","0","C"],"addr":"$cb38","operand1":"B"},"$cb39":{"mnemonic":"SRL","l\
-ength":2,"cycles":[8],"flags":["Z","0","0","C"],"addr":"$cb39","operand1":"C"},"\
-$cb3a":{"mnemonic":"SRL","length":2,"cycles":[8],"flags":["Z","0","0","C"],"addr\
-":"$cb3a","operand1":"D"},"$cb3b":{"mnemonic":"SRL","length":2,"cycles":[8],"fla\
-gs":["Z","0","0","C"],"addr":"$cb3b","operand1":"E"},"$cb3c":{"mnemonic":"SRL","\
-length":2,"cycles":[8],"flags":["Z","0","0","C"],"addr":"$cb3c","operand1":"H"},\
-"$cb3d":{"mnemonic":"SRL","length":2,"cycles":[8],"flags":["Z","0","0","C"],"add\
-r":"$cb3d","operand1":"L"},"$cb3e":{"mnemonic":"SRL","length":2,"cycles":[16],"f\
-lags":["Z","0","0","C"],"addr":"$cb3e","operand1":"(HL)"},"$cb3f":{"mnemonic":"S\
-RL","length":2,"cycles":[8],"flags":["Z","0","0","C"],"addr":"$cb3f","operand1":\
-"A"},"$cb40":{"mnemonic":"BIT","length":2,"cycles":[8],"flags":["Z","0","1","-"]\
-,"addr":"$cb40","operand1":"0","operand2":"B"},"$cb41":{"mnemonic":"BIT","length\
-":2,"cycles":[8],"flags":["Z","0","1","-"],"addr":"$cb41","operand1":"0","operan\
-d2":"C"},"$cb42":{"mnemonic":"BIT","length":2,"cycles":[8],"flags":["Z","0","1",\
-"-"],"addr":"$cb42","operand1":"0","operand2":"D"},"$cb43":{"mnemonic":"BIT","le\
-ngth":2,"cycles":[8],"flags":["Z","0","1","-"],"addr":"$cb43","operand1":"0","op\
-erand2":"E"},"$cb44":{"mnemonic":"BIT","length":2,"cycles":[8],"flags":["Z","0",\
-"1","-"],"addr":"$cb44","operand1":"0","operand2":"H"},"$cb45":{"mnemonic":"BIT"\
-,"length":2,"cycles":[8],"flags":["Z","0","1","-"],"addr":"$cb45","operand1":"0"\
-,"operand2":"L"},"$cb46":{"mnemonic":"BIT","length":2,"cycles":[16],"flags":["Z"\
-,"0","1","-"],"addr":"$cb46","operand1":"0","operand2":"(HL)"},"$cb47":{"mnemoni\
-c":"BIT","length":2,"cycles":[8],"flags":["Z","0","1","-"],"addr":"$cb47","opera\
-nd1":"0","operand2":"A"},"$cb48":{"mnemonic":"BIT","length":2,"cycles":[8],"flag\
-s":["Z","0","1","-"],"addr":"$cb48","operand1":"1","operand2":"B"},"$cb49":{"mne\
-monic":"BIT","length":2,"cycles":[8],"flags":["Z","0","1","-"],"addr":"$cb49","o\
-perand1":"1","operand2":"C"},"$cb4a":{"mnemonic":"BIT","length":2,"cycles":[8],"\
-flags":["Z","0","1","-"],"addr":"$cb4a","operand1":"1","operand2":"D"},"$cb4b":{\
-"mnemonic":"BIT","length":2,"cycles":[8],"flags":["Z","0","1","-"],"addr":"$cb4b\
-","operand1":"1","operand2":"E"},"$cb4c":{"mnemonic":"BIT","length":2,"cycles":[\
-8],"flags":["Z","0","1","-"],"addr":"$cb4c","operand1":"1","operand2":"H"},"$cb4\
-d":{"mnemonic":"BIT","length":2,"cycles":[8],"flags":["Z","0","1","-"],"addr":"$\
-cb4d","operand1":"1","operand2":"L"},"$cb4e":{"mnemonic":"BIT","length":2,"cycle\
-s":[16],"flags":["Z","0","1","-"],"addr":"$cb4e","operand1":"1","operand2":"(HL)\
-"},"$cb4f":{"mnemonic":"BIT","length":2,"cycles":[8],"flags":["Z","0","1","-"],"\
-addr":"$cb4f","operand1":"1","operand2":"A"},"$cb50":{"mnemonic":"BIT","length":\
-2,"cycles":[8],"flags":["Z","0","1","-"],"addr":"$cb50","operand1":"2","operand2\
-":"B"},"$cb51":{"mnemonic":"BIT","length":2,"cycles":[8],"flags":["Z","0","1","-\
-"],"addr":"$cb51","operand1":"2","operand2":"C"},"$cb52":{"mnemonic":"BIT","leng\
-th":2,"cycles":[8],"flags":["Z","0","1","-"],"addr":"$cb52","operand1":"2","oper\
-and2":"D"},"$cb53":{"mnemonic":"BIT","length":2,"cycles":[8],"flags":["Z","0","1\
-","-"],"addr":"$cb53","operand1":"2","operand2":"E"},"$cb54":{"mnemonic":"BIT","\
-length":2,"cycles":[8],"flags":["Z","0","1","-"],"addr":"$cb54","operand1":"2","\
-operand2":"H"},"$cb55":{"mnemonic":"BIT","length":2,"cycles":[8],"flags":["Z","0\
-","1","-"],"addr":"$cb55","operand1":"2","operand2":"L"},"$cb56":{"mnemonic":"BI\
-T","length":2,"cycles":[16],"flags":["Z","0","1","-"],"addr":"$cb56","operand1":\
-"2","operand2":"(HL)"},"$cb57":{"mnemonic":"BIT","length":2,"cycles":[8],"flags"\
-:["Z","0","1","-"],"addr":"$cb57","operand1":"2","operand2":"A"},"$cb58":{"mnemo\
-nic":"BIT","length":2,"cycles":[8],"flags":["Z","0","1","-"],"addr":"$cb58","ope\
-rand1":"3","operand2":"B"},"$cb59":{"mnemonic":"BIT","length":2,"cycles":[8],"fl\
-ags":["Z","0","1","-"],"addr":"$cb59","operand1":"3","operand2":"C"},"$cb5a":{"m\
-nemonic":"BIT","length":2,"cycles":[8],"flags":["Z","0","1","-"],"addr":"$cb5a",\
-"operand1":"3","operand2":"D"},"$cb5b":{"mnemonic":"BIT","length":2,"cycles":[8]\
-,"flags":["Z","0","1","-"],"addr":"$cb5b","operand1":"3","operand2":"E"},"$cb5c"\
-:{"mnemonic":"BIT","length":2,"cycles":[8],"flags":["Z","0","1","-"],"addr":"$cb\
-5c","operand1":"3","operand2":"H"},"$cb5d":{"mnemonic":"BIT","length":2,"cycles"\
-:[8],"flags":["Z","0","1","-"],"addr":"$cb5d","operand1":"3","operand2":"L"},"$c\
-b5e":{"mnemonic":"BIT","length":2,"cycles":[16],"flags":["Z","0","1","-"],"addr"\
-:"$cb5e","operand1":"3","operand2":"(HL)"},"$cb5f":{"mnemonic":"BIT","length":2,\
-"cycles":[8],"flags":["Z","0","1","-"],"addr":"$cb5f","operand1":"3","operand2":\
-"A"},"$cb60":{"mnemonic":"BIT","length":2,"cycles":[8],"flags":["Z","0","1","-"]\
-,"addr":"$cb60","operand1":"4","operand2":"B"},"$cb61":{"mnemonic":"BIT","length\
-":2,"cycles":[8],"flags":["Z","0","1","-"],"addr":"$cb61","operand1":"4","operan\
-d2":"C"},"$cb62":{"mnemonic":"BIT","length":2,"cycles":[8],"flags":["Z","0","1",\
-"-"],"addr":"$cb62","operand1":"4","operand2":"D"},"$cb63":{"mnemonic":"BIT","le\
-ngth":2,"cycles":[8],"flags":["Z","0","1","-"],"addr":"$cb63","operand1":"4","op\
-erand2":"E"},"$cb64":{"mnemonic":"BIT","length":2,"cycles":[8],"flags":["Z","0",\
-"1","-"],"addr":"$cb64","operand1":"4","operand2":"H"},"$cb65":{"mnemonic":"BIT"\
-,"length":2,"cycles":[8],"flags":["Z","0","1","-"],"addr":"$cb65","operand1":"4"\
-,"operand2":"L"},"$cb66":{"mnemonic":"BIT","length":2,"cycles":[16],"flags":["Z"\
-,"0","1","-"],"addr":"$cb66","operand1":"4","operand2":"(HL)"},"$cb67":{"mnemoni\
-c":"BIT","length":2,"cycles":[8],"flags":["Z","0","1","-"],"addr":"$cb67","opera\
-nd1":"4","operand2":"A"},"$cb68":{"mnemonic":"BIT","length":2,"cycles":[8],"flag\
-s":["Z","0","1","-"],"addr":"$cb68","operand1":"5","operand2":"B"},"$cb69":{"mne\
-monic":"BIT","length":2,"cycles":[8],"flags":["Z","0","1","-"],"addr":"$cb69","o\
-perand1":"5","operand2":"C"},"$cb6a":{"mnemonic":"BIT","length":2,"cycles":[8],"\
-flags":["Z","0","1","-"],"addr":"$cb6a","operand1":"5","operand2":"D"},"$cb6b":{\
-"mnemonic":"BIT","length":2,"cycles":[8],"flags":["Z","0","1","-"],"addr":"$cb6b\
-","operand1":"5","operand2":"E"},"$cb6c":{"mnemonic":"BIT","length":2,"cycles":[\
-8],"flags":["Z","0","1","-"],"addr":"$cb6c","operand1":"5","operand2":"H"},"$cb6\
-d":{"mnemonic":"BIT","length":2,"cycles":[8],"flags":["Z","0","1","-"],"addr":"$\
-cb6d","operand1":"5","operand2":"L"},"$cb6e":{"mnemonic":"BIT","length":2,"cycle\
-s":[16],"flags":["Z","0","1","-"],"addr":"$cb6e","operand1":"5","operand2":"(HL)\
-"},"$cb6f":{"mnemonic":"BIT","length":2,"cycles":[8],"flags":["Z","0","1","-"],"\
-addr":"$cb6f","operand1":"5","operand2":"A"},"$cb70":{"mnemonic":"BIT","length":\
-2,"cycles":[8],"flags":["Z","0","1","-"],"addr":"$cb70","operand1":"6","operand2\
-":"B"},"$cb71":{"mnemonic":"BIT","length":2,"cycles":[8],"flags":["Z","0","1","-\
-"],"addr":"$cb71","operand1":"6","operand2":"C"},"$cb72":{"mnemonic":"BIT","leng\
-th":2,"cycles":[8],"flags":["Z","0","1","-"],"addr":"$cb72","operand1":"6","oper\
-and2":"D"},"$cb73":{"mnemonic":"BIT","length":2,"cycles":[8],"flags":["Z","0","1\
-","-"],"addr":"$cb73","operand1":"6","operand2":"E"},"$cb74":{"mnemonic":"BIT","\
-length":2,"cycles":[8],"flags":["Z","0","1","-"],"addr":"$cb74","operand1":"6","\
-operand2":"H"},"$cb75":{"mnemonic":"BIT","length":2,"cycles":[8],"flags":["Z","0\
-","1","-"],"addr":"$cb75","operand1":"6","operand2":"L"},"$cb76":{"mnemonic":"BI\
-T","length":2,"cycles":[16],"flags":["Z","0","1","-"],"addr":"$cb76","operand1":\
-"6","operand2":"(HL)"},"$cb77":{"mnemonic":"BIT","length":2,"cycles":[8],"flags"\
-:["Z","0","1","-"],"addr":"$cb77","operand1":"6","operand2":"A"},"$cb78":{"mnemo\
-nic":"BIT","length":2,"cycles":[8],"flags":["Z","0","1","-"],"addr":"$cb78","ope\
-rand1":"7","operand2":"B"},"$cb79":{"mnemonic":"BIT","length":2,"cycles":[8],"fl\
-ags":["Z","0","1","-"],"addr":"$cb79","operand1":"7","operand2":"C"},"$cb7a":{"m\
-nemonic":"BIT","length":2,"cycles":[8],"flags":["Z","0","1","-"],"addr":"$cb7a",\
-"operand1":"7","operand2":"D"},"$cb7b":{"mnemonic":"BIT","length":2,"cycles":[8]\
-,"flags":["Z","0","1","-"],"addr":"$cb7b","operand1":"7","operand2":"E"},"$cb7c"\
-:{"mnemonic":"BIT","length":2,"cycles":[8],"flags":["Z","0","1","-"],"addr":"$cb\
-7c","operand1":"7","operand2":"H"},"$cb7d":{"mnemonic":"BIT","length":2,"cycles"\
-:[8],"flags":["Z","0","1","-"],"addr":"$cb7d","operand1":"7","operand2":"L"},"$c\
-b7e":{"mnemonic":"BIT","length":2,"cycles":[16],"flags":["Z","0","1","-"],"addr"\
-:"$cb7e","operand1":"7","operand2":"(HL)"},"$cb7f":{"mnemonic":"BIT","length":2,\
-"cycles":[8],"flags":["Z","0","1","-"],"addr":"$cb7f","operand1":"7","operand2":\
-"A"},"$cb80":{"mnemonic":"RES","length":2,"cycles":[8],"flags":["-","-","-","-"]\
-,"addr":"$cb80","operand1":"0","operand2":"B"},"$cb81":{"mnemonic":"RES","length\
-":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cb81","operand1":"0","operan\
-d2":"C"},"$cb82":{"mnemonic":"RES","length":2,"cycles":[8],"flags":["-","-","-",\
-"-"],"addr":"$cb82","operand1":"0","operand2":"D"},"$cb83":{"mnemonic":"RES","le\
-ngth":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cb83","operand1":"0","op\
-erand2":"E"},"$cb84":{"mnemonic":"RES","length":2,"cycles":[8],"flags":["-","-",\
-"-","-"],"addr":"$cb84","operand1":"0","operand2":"H"},"$cb85":{"mnemonic":"RES"\
-,"length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cb85","operand1":"0"\
-,"operand2":"L"},"$cb86":{"mnemonic":"RES","length":2,"cycles":[16],"flags":["-"\
-,"-","-","-"],"addr":"$cb86","operand1":"0","operand2":"(HL)"},"$cb87":{"mnemoni\
-c":"RES","length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cb87","opera\
-nd1":"0","operand2":"A"},"$cb88":{"mnemonic":"RES","length":2,"cycles":[8],"flag\
-s":["-","-","-","-"],"addr":"$cb88","operand1":"1","operand2":"B"},"$cb89":{"mne\
-monic":"RES","length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cb89","o\
-perand1":"1","operand2":"C"},"$cb8a":{"mnemonic":"RES","length":2,"cycles":[8],"\
-flags":["-","-","-","-"],"addr":"$cb8a","operand1":"1","operand2":"D"},"$cb8b":{\
-"mnemonic":"RES","length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cb8b\
-","operand1":"1","operand2":"E"},"$cb8c":{"mnemonic":"RES","length":2,"cycles":[\
-8],"flags":["-","-","-","-"],"addr":"$cb8c","operand1":"1","operand2":"H"},"$cb8\
-d":{"mnemonic":"RES","length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$\
-cb8d","operand1":"1","operand2":"L"},"$cb8e":{"mnemonic":"RES","length":2,"cycle\
-s":[16],"flags":["-","-","-","-"],"addr":"$cb8e","operand1":"1","operand2":"(HL)\
-"},"$cb8f":{"mnemonic":"RES","length":2,"cycles":[8],"flags":["-","-","-","-"],"\
-addr":"$cb8f","operand1":"1","operand2":"A"},"$cb90":{"mnemonic":"RES","length":\
-2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cb90","operand1":"2","operand2\
-":"B"},"$cb91":{"mnemonic":"RES","length":2,"cycles":[8],"flags":["-","-","-","-\
-"],"addr":"$cb91","operand1":"2","operand2":"C"},"$cb92":{"mnemonic":"RES","leng\
-th":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cb92","operand1":"2","oper\
-and2":"D"},"$cb93":{"mnemonic":"RES","length":2,"cycles":[8],"flags":["-","-","-\
-","-"],"addr":"$cb93","operand1":"2","operand2":"E"},"$cb94":{"mnemonic":"RES","\
-length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cb94","operand1":"2","\
-operand2":"H"},"$cb95":{"mnemonic":"RES","length":2,"cycles":[8],"flags":["-","-\
-","-","-"],"addr":"$cb95","operand1":"2","operand2":"L"},"$cb96":{"mnemonic":"RE\
-S","length":2,"cycles":[16],"flags":["-","-","-","-"],"addr":"$cb96","operand1":\
-"2","operand2":"(HL)"},"$cb97":{"mnemonic":"RES","length":2,"cycles":[8],"flags"\
-:["-","-","-","-"],"addr":"$cb97","operand1":"2","operand2":"A"},"$cb98":{"mnemo\
-nic":"RES","length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cb98","ope\
-rand1":"3","operand2":"B"},"$cb99":{"mnemonic":"RES","length":2,"cycles":[8],"fl\
-ags":["-","-","-","-"],"addr":"$cb99","operand1":"3","operand2":"C"},"$cb9a":{"m\
-nemonic":"RES","length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cb9a",\
-"operand1":"3","operand2":"D"},"$cb9b":{"mnemonic":"RES","length":2,"cycles":[8]\
-,"flags":["-","-","-","-"],"addr":"$cb9b","operand1":"3","operand2":"E"},"$cb9c"\
-:{"mnemonic":"RES","length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cb\
-9c","operand1":"3","operand2":"H"},"$cb9d":{"mnemonic":"RES","length":2,"cycles"\
-:[8],"flags":["-","-","-","-"],"addr":"$cb9d","operand1":"3","operand2":"L"},"$c\
-b9e":{"mnemonic":"RES","length":2,"cycles":[16],"flags":["-","-","-","-"],"addr"\
-:"$cb9e","operand1":"3","operand2":"(HL)"},"$cb9f":{"mnemonic":"RES","length":2,\
-"cycles":[8],"flags":["-","-","-","-"],"addr":"$cb9f","operand1":"3","operand2":\
-"A"},"$cba0":{"mnemonic":"RES","length":2,"cycles":[8],"flags":["-","-","-","-"]\
-,"addr":"$cba0","operand1":"4","operand2":"B"},"$cba1":{"mnemonic":"RES","length\
-":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cba1","operand1":"4","operan\
-d2":"C"},"$cba2":{"mnemonic":"RES","length":2,"cycles":[8],"flags":["-","-","-",\
-"-"],"addr":"$cba2","operand1":"4","operand2":"D"},"$cba3":{"mnemonic":"RES","le\
-ngth":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cba3","operand1":"4","op\
-erand2":"E"},"$cba4":{"mnemonic":"RES","length":2,"cycles":[8],"flags":["-","-",\
-"-","-"],"addr":"$cba4","operand1":"4","operand2":"H"},"$cba5":{"mnemonic":"RES"\
-,"length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cba5","operand1":"4"\
-,"operand2":"L"},"$cba6":{"mnemonic":"RES","length":2,"cycles":[16],"flags":["-"\
-,"-","-","-"],"addr":"$cba6","operand1":"4","operand2":"(HL)"},"$cba7":{"mnemoni\
-c":"RES","length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cba7","opera\
-nd1":"4","operand2":"A"},"$cba8":{"mnemonic":"RES","length":2,"cycles":[8],"flag\
-s":["-","-","-","-"],"addr":"$cba8","operand1":"5","operand2":"B"},"$cba9":{"mne\
-monic":"RES","length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cba9","o\
-perand1":"5","operand2":"C"},"$cbaa":{"mnemonic":"RES","length":2,"cycles":[8],"\
-flags":["-","-","-","-"],"addr":"$cbaa","operand1":"5","operand2":"D"},"$cbab":{\
-"mnemonic":"RES","length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cbab\
-","operand1":"5","operand2":"E"},"$cbac":{"mnemonic":"RES","length":2,"cycles":[\
-8],"flags":["-","-","-","-"],"addr":"$cbac","operand1":"5","operand2":"H"},"$cba\
-d":{"mnemonic":"RES","length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$\
-cbad","operand1":"5","operand2":"L"},"$cbae":{"mnemonic":"RES","length":2,"cycle\
-s":[16],"flags":["-","-","-","-"],"addr":"$cbae","operand1":"5","operand2":"(HL)\
-"},"$cbaf":{"mnemonic":"RES","length":2,"cycles":[8],"flags":["-","-","-","-"],"\
-addr":"$cbaf","operand1":"5","operand2":"A"},"$cbb0":{"mnemonic":"RES","length":\
-2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cbb0","operand1":"6","operand2\
-":"B"},"$cbb1":{"mnemonic":"RES","length":2,"cycles":[8],"flags":["-","-","-","-\
-"],"addr":"$cbb1","operand1":"6","operand2":"C"},"$cbb2":{"mnemonic":"RES","leng\
-th":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cbb2","operand1":"6","oper\
-and2":"D"},"$cbb3":{"mnemonic":"RES","length":2,"cycles":[8],"flags":["-","-","-\
-","-"],"addr":"$cbb3","operand1":"6","operand2":"E"},"$cbb4":{"mnemonic":"RES","\
-length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cbb4","operand1":"6","\
-operand2":"H"},"$cbb5":{"mnemonic":"RES","length":2,"cycles":[8],"flags":["-","-\
-","-","-"],"addr":"$cbb5","operand1":"6","operand2":"L"},"$cbb6":{"mnemonic":"RE\
-S","length":2,"cycles":[16],"flags":["-","-","-","-"],"addr":"$cbb6","operand1":\
-"6","operand2":"(HL)"},"$cbb7":{"mnemonic":"RES","length":2,"cycles":[8],"flags"\
-:["-","-","-","-"],"addr":"$cbb7","operand1":"6","operand2":"A"},"$cbb8":{"mnemo\
-nic":"RES","length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cbb8","ope\
-rand1":"7","operand2":"B"},"$cbb9":{"mnemonic":"RES","length":2,"cycles":[8],"fl\
-ags":["-","-","-","-"],"addr":"$cbb9","operand1":"7","operand2":"C"},"$cbba":{"m\
-nemonic":"RES","length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cbba",\
-"operand1":"7","operand2":"D"},"$cbbb":{"mnemonic":"RES","length":2,"cycles":[8]\
-,"flags":["-","-","-","-"],"addr":"$cbbb","operand1":"7","operand2":"E"},"$cbbc"\
-:{"mnemonic":"RES","length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cb\
-bc","operand1":"7","operand2":"H"},"$cbbd":{"mnemonic":"RES","length":2,"cycles"\
-:[8],"flags":["-","-","-","-"],"addr":"$cbbd","operand1":"7","operand2":"L"},"$c\
-bbe":{"mnemonic":"RES","length":2,"cycles":[16],"flags":["-","-","-","-"],"addr"\
-:"$cbbe","operand1":"7","operand2":"(HL)"},"$cbbf":{"mnemonic":"RES","length":2,\
-"cycles":[8],"flags":["-","-","-","-"],"addr":"$cbbf","operand1":"7","operand2":\
-"A"},"$cbc0":{"mnemonic":"SET","length":2,"cycles":[8],"flags":["-","-","-","-"]\
-,"addr":"$cbc0","operand1":"0","operand2":"B"},"$cbc1":{"mnemonic":"SET","length\
-":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cbc1","operand1":"0","operan\
-d2":"C"},"$cbc2":{"mnemonic":"SET","length":2,"cycles":[8],"flags":["-","-","-",\
-"-"],"addr":"$cbc2","operand1":"0","operand2":"D"},"$cbc3":{"mnemonic":"SET","le\
-ngth":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cbc3","operand1":"0","op\
-erand2":"E"},"$cbc4":{"mnemonic":"SET","length":2,"cycles":[8],"flags":["-","-",\
-"-","-"],"addr":"$cbc4","operand1":"0","operand2":"H"},"$cbc5":{"mnemonic":"SET"\
-,"length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cbc5","operand1":"0"\
-,"operand2":"L"},"$cbc6":{"mnemonic":"SET","length":2,"cycles":[16],"flags":["-"\
-,"-","-","-"],"addr":"$cbc6","operand1":"0","operand2":"(HL)"},"$cbc7":{"mnemoni\
-c":"SET","length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cbc7","opera\
-nd1":"0","operand2":"A"},"$cbc8":{"mnemonic":"SET","length":2,"cycles":[8],"flag\
-s":["-","-","-","-"],"addr":"$cbc8","operand1":"1","operand2":"B"},"$cbc9":{"mne\
-monic":"SET","length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cbc9","o\
-perand1":"1","operand2":"C"},"$cbca":{"mnemonic":"SET","length":2,"cycles":[8],"\
-flags":["-","-","-","-"],"addr":"$cbca","operand1":"1","operand2":"D"},"$cbcb":{\
-"mnemonic":"SET","length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cbcb\
-","operand1":"1","operand2":"E"},"$cbcc":{"mnemonic":"SET","length":2,"cycles":[\
-8],"flags":["-","-","-","-"],"addr":"$cbcc","operand1":"1","operand2":"H"},"$cbc\
-d":{"mnemonic":"SET","length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$\
-cbcd","operand1":"1","operand2":"L"},"$cbce":{"mnemonic":"SET","length":2,"cycle\
-s":[16],"flags":["-","-","-","-"],"addr":"$cbce","operand1":"1","operand2":"(HL)\
-"},"$cbcf":{"mnemonic":"SET","length":2,"cycles":[8],"flags":["-","-","-","-"],"\
-addr":"$cbcf","operand1":"1","operand2":"A"},"$cbd0":{"mnemonic":"SET","length":\
-2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cbd0","operand1":"2","operand2\
-":"B"},"$cbd1":{"mnemonic":"SET","length":2,"cycles":[8],"flags":["-","-","-","-\
-"],"addr":"$cbd1","operand1":"2","operand2":"C"},"$cbd2":{"mnemonic":"SET","leng\
-th":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cbd2","operand1":"2","oper\
-and2":"D"},"$cbd3":{"mnemonic":"SET","length":2,"cycles":[8],"flags":["-","-","-\
-","-"],"addr":"$cbd3","operand1":"2","operand2":"E"},"$cbd4":{"mnemonic":"SET","\
-length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cbd4","operand1":"2","\
-operand2":"H"},"$cbd5":{"mnemonic":"SET","length":2,"cycles":[8],"flags":["-","-\
-","-","-"],"addr":"$cbd5","operand1":"2","operand2":"L"},"$cbd6":{"mnemonic":"SE\
-T","length":2,"cycles":[16],"flags":["-","-","-","-"],"addr":"$cbd6","operand1":\
-"2","operand2":"(HL)"},"$cbd7":{"mnemonic":"SET","length":2,"cycles":[8],"flags"\
-:["-","-","-","-"],"addr":"$cbd7","operand1":"2","operand2":"A"},"$cbd8":{"mnemo\
-nic":"SET","length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cbd8","ope\
-rand1":"3","operand2":"B"},"$cbd9":{"mnemonic":"SET","length":2,"cycles":[8],"fl\
-ags":["-","-","-","-"],"addr":"$cbd9","operand1":"3","operand2":"C"},"$cbda":{"m\
-nemonic":"SET","length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cbda",\
-"operand1":"3","operand2":"D"},"$cbdb":{"mnemonic":"SET","length":2,"cycles":[8]\
-,"flags":["-","-","-","-"],"addr":"$cbdb","operand1":"3","operand2":"E"},"$cbdc"\
-:{"mnemonic":"SET","length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cb\
-dc","operand1":"3","operand2":"H"},"$cbdd":{"mnemonic":"SET","length":2,"cycles"\
-:[8],"flags":["-","-","-","-"],"addr":"$cbdd","operand1":"3","operand2":"L"},"$c\
-bde":{"mnemonic":"SET","length":2,"cycles":[16],"flags":["-","-","-","-"],"addr"\
-:"$cbde","operand1":"3","operand2":"(HL)"},"$cbdf":{"mnemonic":"SET","length":2,\
-"cycles":[8],"flags":["-","-","-","-"],"addr":"$cbdf","operand1":"3","operand2":\
-"A"},"$cbe0":{"mnemonic":"SET","length":2,"cycles":[8],"flags":["-","-","-","-"]\
-,"addr":"$cbe0","operand1":"4","operand2":"B"},"$cbe1":{"mnemonic":"SET","length\
-":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cbe1","operand1":"4","operan\
-d2":"C"},"$cbe2":{"mnemonic":"SET","length":2,"cycles":[8],"flags":["-","-","-",\
-"-"],"addr":"$cbe2","operand1":"4","operand2":"D"},"$cbe3":{"mnemonic":"SET","le\
-ngth":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cbe3","operand1":"4","op\
-erand2":"E"},"$cbe4":{"mnemonic":"SET","length":2,"cycles":[8],"flags":["-","-",\
-"-","-"],"addr":"$cbe4","operand1":"4","operand2":"H"},"$cbe5":{"mnemonic":"SET"\
-,"length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cbe5","operand1":"4"\
-,"operand2":"L"},"$cbe6":{"mnemonic":"SET","length":2,"cycles":[16],"flags":["-"\
-,"-","-","-"],"addr":"$cbe6","operand1":"4","operand2":"(HL)"},"$cbe7":{"mnemoni\
-c":"SET","length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cbe7","opera\
-nd1":"4","operand2":"A"},"$cbe8":{"mnemonic":"SET","length":2,"cycles":[8],"flag\
-s":["-","-","-","-"],"addr":"$cbe8","operand1":"5","operand2":"B"},"$cbe9":{"mne\
-monic":"SET","length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cbe9","o\
-perand1":"5","operand2":"C"},"$cbea":{"mnemonic":"SET","length":2,"cycles":[8],"\
-flags":["-","-","-","-"],"addr":"$cbea","operand1":"5","operand2":"D"},"$cbeb":{\
-"mnemonic":"SET","length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cbeb\
-","operand1":"5","operand2":"E"},"$cbec":{"mnemonic":"SET","length":2,"cycles":[\
-8],"flags":["-","-","-","-"],"addr":"$cbec","operand1":"5","operand2":"H"},"$cbe\
-d":{"mnemonic":"SET","length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$\
-cbed","operand1":"5","operand2":"L"},"$cbee":{"mnemonic":"SET","length":2,"cycle\
-s":[16],"flags":["-","-","-","-"],"addr":"$cbee","operand1":"5","operand2":"(HL)\
-"},"$cbef":{"mnemonic":"SET","length":2,"cycles":[8],"flags":["-","-","-","-"],"\
-addr":"$cbef","operand1":"5","operand2":"A"},"$cbf0":{"mnemonic":"SET","length":\
-2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cbf0","operand1":"6","operand2\
-":"B"},"$cbf1":{"mnemonic":"SET","length":2,"cycles":[8],"flags":["-","-","-","-\
-"],"addr":"$cbf1","operand1":"6","operand2":"C"},"$cbf2":{"mnemonic":"SET","leng\
-th":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cbf2","operand1":"6","oper\
-and2":"D"},"$cbf3":{"mnemonic":"SET","length":2,"cycles":[8],"flags":["-","-","-\
-","-"],"addr":"$cbf3","operand1":"6","operand2":"E"},"$cbf4":{"mnemonic":"SET","\
-length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cbf4","operand1":"6","\
-operand2":"H"},"$cbf5":{"mnemonic":"SET","length":2,"cycles":[8],"flags":["-","-\
-","-","-"],"addr":"$cbf5","operand1":"6","operand2":"L"},"$cbf6":{"mnemonic":"SE\
-T","length":2,"cycles":[16],"flags":["-","-","-","-"],"addr":"$cbf6","operand1":\
-"6","operand2":"(HL)"},"$cbf7":{"mnemonic":"SET","length":2,"cycles":[8],"flags"\
-:["-","-","-","-"],"addr":"$cbf7","operand1":"6","operand2":"A"},"$cbf8":{"mnemo\
-nic":"SET","length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cbf8","ope\
-rand1":"7","operand2":"B"},"$cbf9":{"mnemonic":"SET","length":2,"cycles":[8],"fl\
-ags":["-","-","-","-"],"addr":"$cbf9","operand1":"7","operand2":"C"},"$cbfa":{"m\
-nemonic":"SET","length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cbfa",\
-"operand1":"7","operand2":"D"},"$cbfb":{"mnemonic":"SET","length":2,"cycles":[8]\
-,"flags":["-","-","-","-"],"addr":"$cbfb","operand1":"7","operand2":"E"},"$cbfc"\
-:{"mnemonic":"SET","length":2,"cycles":[8],"flags":["-","-","-","-"],"addr":"$cb\
-fc","operand1":"7","operand2":"H"},"$cbfd":{"mnemonic":"SET","length":2,"cycles"\
-:[8],"flags":["-","-","-","-"],"addr":"$cbfd","operand1":"7","operand2":"L"},"$c\
-bfe":{"mnemonic":"SET","length":2,"cycles":[16],"flags":["-","-","-","-"],"addr"\
-:"$cbfe","operand1":"7","operand2":"(HL)"},"$cbff":{"mnemonic":"SET","length":2,\
-"cycles":[8],"flags":["-","-","-","-"],"addr":"$cbff","operand1":"7","operand2":\
-"A"}}
-"""
