--- gbasm/global.py
@@ -1,13 +1,0 @@
-"""
-Global storage for the Gamy Boy Assembler
-"""
-
-from singleton_decorator import singleton
-from gbasm.label import Label, Labels
-
-@singleton
-class Global:
-
-    def __init__(self):
-        self._labels = Labels()
-        self._global_labels = Labels()
