--- gbasm/instruction/__init__.py
@@ -1,16 +1,0 @@
-"""
-Game Boy Assembler Instruction related classes
-"""
-
-from gbasm.instruction.registers import Registers
-from gbasm.instruction.instruction_set import InstructionSet
-from gbasm.instruction.instruction import Instruction
-from gbasm.instruction.instruction_pointer import InstructionPointer
-from gbasm.instruction.lexer_parser import LexerResults, LexerTokens
-from gbasm.instruction.lexer_parser import InstructionParser
-
-__all__ = [
-    "registers", "instruction_set", "instruction", "instruction_pointer",
-    "lexer_parser"
-]
-
